Add files via upload
The goal of this project is to analyze the Google Play Store apps dataset, focusing on categories, ratings, sizes, and other relevant information. Additionally, customer reviews of Android apps are included for analysis. The objective is to identify critical factors that influence app engagement and overall success using Python tools such as numpy, pandas, and matplotlib.